Collapsing solnet.util
