+++ src/Solnet.Examples/Solnet.Examples.csproj
@@ -11,7 +11,6 @@
       <ProjectReference Include="..\Solnet.Programs\Solnet.Programs.csproj" />
       <ProjectReference Include="..\Solnet.Rpc\Solnet.Rpc.csproj" />
       <ProjectReference Include="..\Solnet.Models\Solnet.Models.csproj" />
-      <ProjectReference Include="..\Solnet.Util\Solnet.Util.csproj" />
       <ProjectReference Include="..\Solnet.Wallet\Solnet.Wallet.csproj" />
     </ItemGroup>
