+++ src/Solnet.KeyStore/Model/CryptoInfo.cs
@@ -1,5 +1,4 @@
 using System.Text.Json.Serialization;
-using Solnet.Util;
 
 namespace Solnet.KeyStore.Model
 {
