+++ src/Solnet.KeyStore/Solnet.KeyStore.csproj
@@ -12,7 +12,6 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Solnet.Wallet\Solnet.Wallet.csproj" />
-      <ProjectReference Include="..\Solnet.Util\Solnet.Util.csproj" />
     </ItemGroup>
 
 </Project>
