--- src/Solnet.Util/Solnet.Util.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="NBitcoin" Version="5.0.77" />
-      <PackageReference Include="Portable.BouncyCastle" Version="1.8.2" />
-    </ItemGroup>
-</Project>
