+++ src/Solnet.Wallet/Solnet.Wallet.csproj
@@ -15,7 +15,6 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Solnet.Models\Solnet.Models.csproj" />
-      <ProjectReference Include="..\Solnet.Util\Solnet.Util.csproj" />
     </ItemGroup>
     
 </Project>
