+++ test/Solnet.Wallet.Test/AccountTest.cs
@@ -1,6 +1,5 @@
 using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Solnet.Util;
 
 namespace Solnet.Wallet.Test
 {
