+++ test/Solnet.Wallet.Test/WalletTest.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NBitcoin;
-using Solnet.Util;
 
 namespace Solnet.Wallet.Test
 {
